Moved tests into a subdirectory
